requirements.txt
bitsandbytes==0.43.0
peft==0.8.2
trl==0.7.10
accelerate==0.27.1
datasets==2.17.0
transformers==4.38.0
python-dotenv